requirements.txt
aniso8601==9.0.1
anyio==3.4.0
asgiref==3.4.1
attrs==21.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
clickclick==20.10.2
connexion==2.9.0
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.6
fastapi==0.70.1
Flask==1.1.4
Flask-RESTful==0.3.9
flask-restx==0.5.1
h11==0.12.0
httptools==0.3.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.2
inflection==0.5.1
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pydantic==1.8.2
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-order==1.0.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
swagger-ui-bundle==0.0.9
toml==0.10.2
tomli==1.2.3
typed-ast==1.5.1
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
Werkzeug==1.0.1
zipp==3.6.0
